Wed, Jan 3, 2024, 11:28 PM +08:00

+++ readability/readability.c
@@ -21,7 +21,7 @@ int main(void)
     // Compute the Coleman-Liau index
     int L = ((float) letters / (float) words) * 100;
     int S = ((float) sentences / (float) words) * 100;
-    int index = 0.0588 * ((float) L - 0.296) * ((float) S - 15.8);
+    float index = 0.0588 * ((float) L - 0.296) * ((float) S - 15.8);
 
     // Print the grade level
     if ((int) round(index) >= 16)
